feat(validators): add validators to requests
