+++ src/application/user/commands/update_user.py
@@ -6,7 +6,7 @@ from didiator import EventMediator
 from src.application.common.command import Command, CommandHandler
 from src.application.common.interfaces.mapper import Mapper
 from src.application.common.interfaces.uow import UnitOfWork
-from src.application.user import dto
+from src.application.user import dto, validators
 from src.application.user.interfaces import UserRepo
 from src.domain.base.constants import Empty
 from src.domain.user.value_objects import UserId, Username
@@ -17,6 +17,10 @@ class UpdateUserData:
     username: str | Empty = Empty.UNSET
     first_name: str | Empty = Empty.UNSET
     last_name: str | None | Empty = Empty.UNSET
+
+    def __post_init__(self) -> None:
+        if self.username is not Empty.UNSET:
+            validators.validate_username(self.username)
 
 
 @dataclass(frozen=True)
