+++ src/application/user/validators/__init__.py
@@ -1,0 +1,5 @@
+from .username import validate_username
+
+__all__ = (
+    "validate_username",
+)
